Refactor out taiko Peaks skill
